requirements.txt
path.py
subliminal
